refactor: move secret-ref schema stripping to core.schema.check()
Instead of bypassing plugin_obj.check_schema() entirely when secret
refs are detected, move the strip logic into core.schema.check().
This preserves all custom code validation in each plugin's
check_schema (regex_uri compilation, headers checks, group field
validation, claim_schema, etc.) while only bypassing jsonschema
constraints for the specific fields containing secret references.
