+++ apisix/plugin.lua
@@ -31,7 +31,6 @@ local crc32         = ngx.crc32_short
 local ngx_exit      = ngx.exit
 local pkg_loaded    = package.loaded
 local sort_tab      = table.sort
-local tab_remove    = table.remove
 local pcall         = pcall
 local ipairs        = ipairs
 local pairs         = pairs
@@ -964,73 +963,6 @@ function _M.conf_version(conf)
 end
 
 
-local function strip_secret_refs(conf, schema)
-    if type(conf) ~= "table" or type(schema) ~= "table" then
-        return
-    end
-
-    local props = schema.properties
-    local removed = {}
-
-    for k, v in pairs(conf) do
-        if type(v) == "string" and secret.is_secret_ref(v) then
-            conf[k] = nil
-            core.table.insert(removed, k)
-            if props then
-                props[k] = nil
-            end
-        elseif type(v) == "table" then
-            local sub_schema = props and props[k]
-            if sub_schema then
-                if sub_schema.type == "object" or sub_schema.properties then
-                    strip_secret_refs(v, sub_schema)
-                elseif sub_schema.type == "array" and sub_schema.items then
-                    if sub_schema.items.type == "string" then
-                        -- strip secret refs from string arrays in-place
-                        for i = #v, 1, -1 do
-                            if type(v[i]) == "string" and secret.is_secret_ref(v[i]) then
-                                tab_remove(v, i)
-                            end
-                        end
-                    else
-                        for _, item in ipairs(v) do
-                            if type(item) == "table" then
-                                strip_secret_refs(item, sub_schema.items)
-                            end
-                        end
-                    end
-                end
-            end
-        end
-    end
-
-    if #removed > 0 then
-        local function strip_required(s)
-            if s.required then
-                local new_req = {}
-                for _, r in ipairs(s.required) do
-                    local keep = true
-                    for _, rm in ipairs(removed) do
-                        if r == rm then keep = false; break end
-                    end
-                    if keep then
-                        core.table.insert(new_req, r)
-                    end
-                end
-                s.required = #new_req > 0 and new_req or nil
-            end
-            for _, kw in ipairs({"allOf", "anyOf", "oneOf"}) do
-                if s[kw] then
-                    for _, sub in ipairs(s[kw]) do
-                        strip_required(sub)
-                    end
-                end
-            end
-        end
-        strip_required(schema)
-    end
-end
-
 
 local function check_single_plugin_schema(name, plugin_conf, schema_type, skip_disabled_plugin)
     if type(plugin_conf) ~= "table" then
@@ -1051,31 +983,7 @@ local function check_single_plugin_schema(name, plugin_conf, schema_type, skip_d
     end
 
     if plugin_obj.check_schema then
-        local ok, err
-
-        if secret.has_secret_ref(plugin_conf) then
-            -- Strip secret ref fields recursively so they bypass all schema
-            -- constraints (enum, pattern, minLength, maxLength, etc.).
-            -- We deep-copy both conf and schema, remove secret-ref leaves from
-            -- the conf copy, and remove the corresponding property definitions
-            -- and required entries from the schema copy.
-            local conf_copy = core.table.deepcopy(plugin_conf)
-            local schema_raw
-            if schema_type == _M.TYPE_CONSUMER then
-                schema_raw = plugin_obj.consumer_schema or plugin_obj.schema
-            elseif schema_type == _M.TYPE_METADATA then
-                schema_raw = plugin_obj.metadata_schema
-            else
-                schema_raw = plugin_obj.schema
-            end
-            local schema_copy = core.table.deepcopy(schema_raw)
-
-            strip_secret_refs(conf_copy, schema_copy)
-            ok, err = core.schema.check(schema_copy, conf_copy)
-        else
-            ok, err = plugin_obj.check_schema(plugin_conf, schema_type)
-        end
-
+        local ok, err = plugin_obj.check_schema(plugin_conf, schema_type)
         if not ok then
             return false, "failed to check the configuration of plugin "
                 .. name .. " err: " .. err
@@ -1307,15 +1215,7 @@ local function stream_check_schema(plugins_conf, schema_type, skip_disabled_plug
         end
 
         if plugin_obj.check_schema then
-            local ok, err
-            if secret.has_secret_ref(plugin_conf) then
-                local conf_copy = core.table.deepcopy(plugin_conf)
-                local schema_copy = core.table.deepcopy(plugin_obj.schema)
-                strip_secret_refs(conf_copy, schema_copy)
-                ok, err = core.schema.check(schema_copy, conf_copy)
-            else
-                ok, err = plugin_obj.check_schema(plugin_conf, schema_type)
-            end
+            local ok, err = plugin_obj.check_schema(plugin_conf, schema_type)
             if not ok then
                 return false, "failed to check the configuration of "
                               .. "stream plugin [" .. name .. "]: " .. err
